Refactor AuthenticatedGuard imports and enhance configurations: - Moved `AuthenticatedGuard` import paths to `utils/guards/local.guard` across all controllers for better organization. - Added `SubscriptionsModule` and global `ConfigModule` in `AppModule` for enhanced modularization and configuration management. - Updated `User` entity: removed relationships with `Country`, `Definition`, and `Word` while adding a `stripeCustomerId` column.
